Refactor for 23a Python Importer demo
